task_manager_project/requirements.txt
fastapi
uvicorn
tortoise-orm
redis
pyjwt
passlib[bcrypt]
streamlit
requests
